# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Название вашего API",
        "description": "Описание вашего API",
        "contact": {
            "name": "Ваше имя",
            "email": "ваш.email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/accordions": {
            "get": {
                "tags": [
                    "Accordions"
                ],
                "summary": "Get list of accordions",
                "operationId": "34142635212e590e8bb786c3e6bcc761",
                "responses": {
                    "200": {
                        "description": "List of locations"
                    }
                }
            },
            "post": {
                "tags": [
                    "Accordions"
                ],
                "summary": "Create a new accordion",
                "operationId": "eeeb6adaefae38577b60430e8e5ed416",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address": {
                                        "type": "string"
                                    },
                                    "tel": {
                                        "type": "number"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Accordion created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/accordions/{id}": {
            "get": {
                "tags": [
                    "Accordions"
                ],
                "summary": "Get a specific location",
                "operationId": "f35c34b68fe7ada01dcb49f12b722c15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Accordion ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Accordion details"
                    },
                    "404": {
                        "description": "Accordion not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Accordions"
                ],
                "summary": "Update a specific accordion",
                "operationId": "7168d48c57895084befead623e4af396",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Accordion ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address": {
                                        "type": "string"
                                    },
                                    "tel": {
                                        "type": "number"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "location": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Accordion updated successfully"
                    },
                    "404": {
                        "description": "Accordion not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Accordions"
                ],
                "summary": "Delete a specific accordion",
                "operationId": "fbe82d5a2ed25764765928b225505e20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Accordion ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Accordion deleted successfully"
                    },
                    "404": {
                        "description": "Accordion not found"
                    }
                }
            }
        },
        "/api/banners": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get list of banners",
                "operationId": "9f718c7db0d13e4b5be26b80595f4cf1",
                "responses": {
                    "200": {
                        "description": "List of banners"
                    }
                }
            },
            "post": {
                "tags": [
                    "Banners"
                ],
                "summary": "Store a new banner",
                "operationId": "546707dc3442c51e694ec312df3664e9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "text",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Banner Title"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Banner Text"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Banner created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/banners/{id}": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get details of a specific banner",
                "operationId": "13d01a36bda0d8b491a382619c6bc469",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner details"
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Banners"
                ],
                "summary": "Update details of a specific banner",
                "operationId": "89bb1eb70cfa7909040b083ebe79f6df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the banner",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "text",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Banner Title"
                                    },
                                    "text": {
                                        "type": "string",
                                        "example": "Updated Banner Description"
                                    },
                                    "image": {
                                        "type": "string",
                                        "example": "Image Title"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Banner updated successfully"
                    },
                    "404": {
                        "description": "Banner not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Banners"
                ],
                "summary": "Delete a banner",
                "operationId": "00a27f787ec0124f0b98cfdeecf6090a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the banner to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Banner deleted successfully"
                    },
                    "404": {
                        "description": "Banner not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}